+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Added
 
+- `RDF.Graph.clear_name/1`
 - `RDF.List.prepend/3`
 - `RDF.Diff.invert/1`
 - Some performance improvements on the Turtle and TriG encoder